# Don't load block list in ...Truncated methods