@AudriusButkevicius 2015-10-21
Are you saying nobody uses FileTruncated.Size()?