@calmh 2015-10-21
No, I'm saying we had the CachedSize field ready and waiting in FileInfo but apparently wasn't using it, so I reused that and let the original FileInfo.Size() method handle it.